requirements.txt
pip install -r requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pandas>=1.0.0
openpyxl>=3.0.0       # pandas 读写 Excel
Pillow>=7.0.0
tqdm>=4.36.0
